requirements.txt
PySide6>=6.6
biopython>=1.83
PyYAML>=6.0.1
jsonschema>=4.21
